+++ bogenliga/bogenliga-business-logic/src/test/java/de/bogenliga/application/business/schuetzenstatistikLetzeJahre/impl/business/SchuetzenstatistikLetzteJahreImplTest.java
@@ -28,12 +28,12 @@ public class SchuetzenstatistikLetzteJahreImplTest {
     private static final long sportjahr = (long) 2002;
 
     private static final String schuetzenname = "Name Schütze";
-    private static final long sportjahr1 = (long) 8;
-    private static final long sportjahr2 = (long) 7;
-    private static final long sportjahr3 = (long) 6.5;
-    private static final long sportjahr4 = (long) 7.8;
-    private static final long sportjahr5 = (long) 8.2;
-    private static final long allejahre_schnitt = (long) 7.5;
+    private static final float sportjahr1 = (float) 8;
+    private static final float sportjahr2 = (float) 7;
+    private static final float sportjahr3 = (float) 6.5;
+    private static final float sportjahr4 = (float) 7.8;
+    private static final float sportjahr5 = (float) 8.2;
+    private static final float allejahre_schnitt = (float) 7.5;
 
     @Rule
     public MockitoRule mockitoRule = MockitoJUnit.rule();
